fix: remove prism emoji and restore beamx arrows
