+++ src/canvas/prism.rs
@@ -1,12 +1,6 @@
-use ratatui::{backend::Backend, layout::Rect, widgets::Paragraph, style::Style, Frame};
+use ratatui::{backend::Backend, layout::Rect, Frame};
 
 /// Render a small PrismX emblem in the top-right corner of the canvas.
-pub fn render_prism<B: Backend>(f: &mut Frame<B>, area: Rect) {
-    if area.width < 2 || area.height < 1 {
-        return;
-    }
-    let x = area.right().saturating_sub(2);
-    let y = area.y;
-    let style = Style::default();
-    f.render_widget(Paragraph::new("🔷").style(style), Rect::new(x, y, 2, 1));
+pub fn render_prism<B: Backend>(_: &mut Frame<B>, _: Rect) {
+    // Icon rendering removed – previously drew a debug prism glyph
 }
