+++ src/ui/mod.rs
@@ -1,3 +1,4 @@
 pub mod beamx;
 pub mod shortcuts;
 pub mod components;
+pub mod render;
