+++ tests/golden/gemx.snapshot
@@ -1,7 +1,7 @@
 ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━           ┓
-┃[A] Auto-Arrange                                 
-┃ Node A                                          
-┃                                                 
+┃[A] Auto-Arrange
+┃ Node A
+┃
 ┃             Node B                             ┃
 ┃                             ╭──────────╮       ┃
 ┃                             │Zoom: 0.5x│       ┃
